feat: database seeder migration payment methods

+++ database/migrations/2024_12_18_081315_create_payment_methods_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('payment_methods', function (Blueprint $table) {
           $table->id();
           $table->string('name');
-          $table->integer('account_number');
+          $table->bigInteger('account_number');
           $table->string('image')->nullable();
           $table->timestamps();
         });
